pyproject.toml
[project]
name = "openems-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "numpy>=2.2.2",
    "matplotlib>=3.10.0",
    "h5py>=3.13.0",
    "ruff>=0.9.5",
    "scipy>=1.15.1",
    "tqdm>=4.67.1",
    "typer-slim>=0.15.2",
]

[project.optional-dependencies]
jax = ["jax[cuda12]==0.6.1", "flax>=0.10.6"]
# jax = ["jax[cuda12]==0.5.3", "flax>=0.10.6"]
# torch = ["torch>=2.7.0", "torchvision>=0.22.0"]

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pyflakes
    "F",
]
exclude = ["src/*"] # these scripts run inside OpenEMS container
